+++ lib/builder-webpack4/src/preview/iframe-webpack.config.ts
@@ -98,6 +98,7 @@ export default async (options: Options & Record<string, any>): Promise<Configura
     const storiesPath = path.resolve(path.join(workingDir, storiesFilename));
 
     virtualModuleMapping[storiesPath] = toImportFn(stories);
+
     const configEntryPath = path.resolve(path.join(workingDir, 'storybook-config-entry.js'));
     virtualModuleMapping[configEntryPath] = handlebars(
       await readTemplate(path.join(__dirname, 'virtualModuleModernEntry.js.handlebars')),
@@ -105,7 +106,8 @@ export default async (options: Options & Record<string, any>): Promise<Configura
         storiesFilename,
         configs,
       }
-    );
+      // We need to double escape `\` for webpack. We may have some in windows paths
+    ).replace(/\\/g, '\\\\');
     entries.push(configEntryPath);
   } else {
     const frameworkInitEntry = path.resolve(