+++ lib/core-server/src/utils/StoryIndexGenerator.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 import fs from 'fs-extra';
 import glob from 'globby';
+import slash from 'slash';
 
 import {
   autoTitleFromSpecifier,
@@ -57,7 +58,9 @@ export class StoryIndexGenerator {
       this.specifiers.map(async (specifier) => {
         const pathToSubIndex = {} as SpecifierStoriesCache;
 
-        const fullGlob = path.join(this.options.workingDir, specifier.directory, specifier.files);
+        const fullGlob = slash(
+          path.join(this.options.workingDir, specifier.directory, specifier.files)
+        );
         const files = await glob(fullGlob);
         files.forEach((absolutePath: Path) => {
           const ext = path.extname(absolutePath);
@@ -100,7 +103,7 @@ export class StoryIndexGenerator {
       const entry = this.storyIndexEntries.get(specifier);
       const fileStories = {} as StoryIndex['stories'];
 
-      const importPath = relativePath[0] === '.' ? relativePath : `./${relativePath}`;
+      const importPath = slash(relativePath[0] === '.' ? relativePath : `./${relativePath}`);
       const defaultTitle = autoTitleFromSpecifier(importPath, specifier);
       const csf = (await readCsfOrMdx(absolutePath, { defaultTitle })).parse();
       csf.stories.forEach(({ id, name }) => {